% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_cjsg_tjms.R
\name{baixar_cjsg_tjms}
\alias{baixar_cjsg_tjms}
\title{Baixa consulta jurisprudencial do tjms}
\usage{
baixar_cjsg_tjms(
  livre = "",
  aspas = FALSE,
  classe = "",
  assunto = "",
  orgao_julgador = "",
  inicio = "",
  fim = "",
  tipo = "A",
  diretorio = "."
)
}
\arguments{
\item{livre}{palavra ou texto a ser buscado nas ementas e nos acórdãos}

\item{aspas}{lógico. Colocar a expressão entre aspas?}

\item{classe}{Código da classe processual}

\item{assunto}{Código do assunto}

\item{orgao_julgador}{Órgão julgador}

\item{inicio}{data inicial}

\item{fim}{Data final}

\item{tipo}{"A" Para acórdãos, "D" para decisões monocráticas.}

\item{diretorio}{Diretório onde serão armazenadas as páginas html.}
}
\value{
baixa os htmls das decisões de segunda instância
}
\description{
Baixa consulta jurisprudencial do tjms
}
\examples{
\dontrun{
baixar_cjsg(livre = "Lei Maria da Penha")
}

}
\keyword{tjms,acórdãos}
