% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_cpopg_tjms.R
\name{baixar_cpopg_tjms}
\alias{baixar_cpopg_tjms}
\title{Baixa consulta processual de primeiro grau}
\usage{
baixar_cpopg_tjms(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{Número do processo (cnj)}

\item{diretorio}{Diretório onde serão armazenados os htmls}
}
\value{
html com dados processuais
}
\description{
Baixa consulta processual de primeiro grau
}
