% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_sentencas_tjms.R
\name{baixar_sentencas_tjms}
\alias{baixar_sentencas_tjms}
\title{Baixa sentencas do tjms a partir de tabela de sentencas retirada da
    movimentação}
\usage{
baixar_sentencas_tjms(df, diretorio = ".")
}
\arguments{
\item{df}{Movimentação com url da sentença}

\item{diretorio}{diretório onde serão armazenadas as sentenças}
}
\value{
Arquivos sem extenção. Não dá para saber se o arquivo é pdf ou
    rtf.
}
\description{
Baixa sentencas do tjms a partir de tabela de sentencas retirada da
    movimentação
}
\details{
Os arquivos são armazenados sem extenção, pois não sabemos se
    eles são rtf ou pdf. A função ler_sentencas_tjms se encarrega de
    fazer essa verificação.
}
