% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_cjsg_tjms.R
\name{ler_cjsg_tjms}
\alias{ler_cjsg_tjms}
\title{Ler metadados das decis천es do TJMS}
\usage{
ler_cjsg_tjms(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquvos a serem lidos}

\item{diretorio}{Diret처rio onde se encontram os htmls, caso n찾o informe os arquivos}
}
\value{
tabela com metadados jurisprudenciais.
}
\description{
Ler metadados das decis천es do TJMS
}
\examples{
\dontrun{
cjsg <- ler_cjsg_tjms()
}

}
