% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados_cpopg_tjms.R
\name{ler_dados_cpopg_tjms}
\alias{ler_dados_cpopg_tjms}
\title{Lê metadados dos processos de primeiro grau do TJMS}
\usage{
ler_dados_cpopg_tjms(arquivos = NULL, diretorio = ".", wide = FALSE)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Se não informados os arquivos, informar o diretório}

\item{wide}{Se TRUE, as informações serão dispostas em wide format.}
}
\value{
tibble
}
\description{
Lê metadados dos processos de primeiro grau do TJMS
}
