% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_movimentacao_tjms.R
\name{ler_movimentacao}
\alias{ler_movimentacao}
\title{Ler movimentacao tjms}
\usage{
ler_movimentacao(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor com os caminhos para os arquivos}

\item{diretorio}{Se não informar os arquivos, informar o diretório}
}
\value{
tibble
}
\description{
Ler movimentacao tjms
}
