% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_movimentacao_tjms.R
\name{ler_movimentacao_tjms}
\alias{ler_movimentacao_tjms}
\title{Lê movimentação processual do tjms}
\usage{
ler_movimentacao_tjms(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor com os caminhos para os arquivos}

\item{diretorio}{Se não informar os arquivos, informar o diretório}
}
\value{
tibble
}
\description{
Lê movimentação processual do tjms
}
