% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_sentencas_tjms.R
\name{ler_sentencas_tjms}
\alias{ler_sentencas_tjms}
\title{Lê sentenças baixadas com função baixar_sentencas_tjms}
\usage{
ler_sentencas_tjms(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Se não informados os arquivos, informar o diretório}
}
\value{
tibble com número do processo, número do documento e inteiro teor.
}
\description{
Lê sentenças baixadas com função baixar_sentencas_tjms
}
\details{
Esta função antes de ler, verifica se o texto é rtf ou pdf.
}
