% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjms_baixar_cpopg.R
\name{tjms_baixar_cpopg}
\alias{tjms_baixar_cpopg}
\title{Baixa consulta processual de primeiro grau}
\usage{
tjms_baixar_cpopg(processos = NULL, diretorio = ".")
}
\arguments{
\item{processos}{Número do processo (cnj)}

\item{diretorio}{Diretório onde serão armazenados os htmls}
}
\value{
html com dados processuais
}
\description{
Baixa consulta processual de primeiro grau
}
