% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjms_combinar_docs.R
\name{tjms_combinar_docs}
\alias{tjms_combinar_docs}
\title{Combinar pdfs baixados}
\usage{
tjms_combinar_docs(arquivos = NULL, dir_origem = ".", dir_destino = NULL)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{dir_origem}{Diretório de origem se arg arquivos não informado}

\item{dir_destino}{Diretório destino dos pdfs combinados}
}
\value{
pdf
}
\description{
Combinar pdfs baixados
}
