% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjms_ler_cjsg.R
\name{tjms_ler_cjsg}
\alias{tjms_ler_cjsg}
\title{Ler metadados das decis천es do TJMS}
\usage{
tjms_ler_cjsg(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquvos a serem lidos}

\item{diretorio}{Diret처rio onde se encontram os htmls, caso n찾o informe os arquivos}
}
\value{
tabela com metadados jurisprudenciais.
}
\description{
Ler metadados das decis천es do TJMS
}
\examples{
\dontrun{
cjsg <- ler_cjsg_tjms()
}

}
