% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjms_ler_dados_cpopg.R
\name{tjms_ler_dados_cpopg}
\alias{tjms_ler_dados_cpopg}
\title{Lê metadados dos processos de primeiro grau do TJMS}
\usage{
tjms_ler_dados_cpopg(arquivos = NULL, diretorio = ".", wide = FALSE)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Se não informados os arquivos, informar o diretório}

\item{wide}{Se TRUE, as informações serão dispostas em wide format.}
}
\value{
tibble
}
\description{
Lê metadados dos processos de primeiro grau do TJMS
}
