% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjms_ler_tabela_docs.R
\name{tjms_ler_tabela_docs}
\alias{tjms_ler_tabela_docs}
\title{Ler tabelas baixadas com tjms_baixar_tabela_docs}
\usage{
tjms_ler_tabela_docs(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se não informar diretório}
}
\value{
tibble
}
\description{
Ler tabelas baixadas com tjms_baixar_tabela_docs
}
